+++ torchy_baselines/sac/policies.py
@@ -1,4 +1,4 @@
-from typing import Optional, List, Tuple, Callable, Union
+from typing import Optional, List, Tuple, Callable, Union, Type
 
 import gym
 import torch as th
@@ -24,7 +24,7 @@ class Actor(BasePolicy):
     :param features_extractor: (nn.Module) Network to extract features
         (a CNN when using images, a nn.Flatten() layer otherwise)
     :param features_dim: (int) Number of features
-    :param activation_fn: (nn.Module) Activation function
+    :param activation_fn: (Type[nn.Module]) Activation function
     :param use_sde: (bool) Whether to use State Dependent Exploration or not
     :param log_std_init: (float) Initial value for the log standard deviation
     :param full_std: (bool) Whether to use (n_features x n_actions) parameters
@@ -44,7 +44,7 @@ class Actor(BasePolicy):
                  net_arch: List[int],
                  features_extractor: nn.Module,
                  features_dim: int,
-                 activation_fn: nn.Module = nn.ReLU,
+                 activation_fn: Type[nn.Module] = nn.ReLU,
                  use_sde: bool = False,
                  log_std_init: float = -3,
                  full_std: bool = True,
@@ -151,7 +151,7 @@ class Critic(BasePolicy):
     :param features_extractor: (nn.Module) Network to extract features
         (a CNN when using images, a nn.Flatten() layer otherwise)
     :param features_dim: (int) Number of features
-    :param activation_fn: (nn.Module) Activation function
+    :param activation_fn: (Type[nn.Module]) Activation function
     :param normalize_images: (bool) Whether to normalize images or not,
          dividing by 255.0 (True by default)
     """
@@ -160,7 +160,7 @@ class Critic(BasePolicy):
                  net_arch: List[int],
                  features_extractor: nn.Module,
                  features_dim: int,
-                 activation_fn: nn.Module = nn.ReLU,
+                 activation_fn: Type[nn.Module] = nn.ReLU,
                  normalize_images: bool = True):
         super(Critic, self).__init__(observation_space, action_space,
                                      features_extractor=features_extractor,
@@ -191,7 +191,7 @@ class SACPolicy(BasePolicy):
     :param lr_schedule: (callable) Learning rate schedule (could be constant)
     :param net_arch: (Optional[List[int]]) The specification of the policy and value networks.
     :param device: (str or th.device) Device on which the code should run.
-    :param activation_fn: (nn.Module) Activation function
+    :param activation_fn: (Type[nn.Module]) Activation function
     :param use_sde: (bool) Whether to use State Dependent Exploration or not
     :param log_std_init: (float) Initial value for the log standard deviation
     :param sde_net_arch: ([int]) Network architecture for extracting features
@@ -209,7 +209,7 @@ class SACPolicy(BasePolicy):
                  lr_schedule: Callable,
                  net_arch: Optional[List[int]] = None,
                  device: Union[th.device, str] = 'cpu',
-                 activation_fn: nn.Module = nn.ReLU,
+                 activation_fn: Type[nn.Module] = nn.ReLU,
                  use_sde: bool = False,
                  log_std_init: float = -3,
                  sde_net_arch: Optional[List[int]] = None,